+++ src/main/kotlin/com/academia/smartgym/domain/repository/ExercicioRepository.kt
@@ -4,9 +4,9 @@ import com.academia.smartgym.domain.model.Exercicio
 
 interface ExercicioRepository {
     fun findAll(): List<Exercicio>
-    fun findById(id: Long): Exercicio?
+    fun findById(id: Int): Exercicio?
     fun save(exercicio: Exercicio): Exercicio
-    fun deleteById(id: Long)
-    fun findByMaquinaId(maquinaId: Long): List<Exercicio>
+    fun deleteById(id: Int)
+    fun findByMaquinaId(maquinaId: Int): List<Exercicio>
     fun findByNomeContainingIgnoreCase(nome: String): List<Exercicio>
 }
